make topPod as usage percentage

+++ pkg/metrics/provider.go
@@ -4,30 +4,30 @@ import "github.com/smvfal/faas-monitor/pkg/types"
 
 type Provider interface {
 
-	// get the names of the deployed functions
+	// Functions provides the names of the deployed functions
 	Functions() ([]string, error)
 
-	// get the function replicas' number
+	// FunctionReplicas provides the function replicas' number
 	FunctionReplicas(functionName string) (int, error)
 
-	// get the function invocation rate
+	// FunctionInvocationRate provides the function invocation rate
 	FunctionInvocationRate(functionName string, sinceSeconds int64) (float64, error)
 
-	// get function's average response time
+	// ResponseTime provides function's average response time
 	ResponseTime(functionName string, sinceSeconds int64) (float64, error)
 
-	// get function's average processing time
+	// ProcessingTime provides function's average processing time
 	ProcessingTime(functionName string, sinceSeconds int64) (float64, error)
 
-	// get function's throughput
+	// Throughput provides function's throughput
 	Throughput(functionName string, sinceSeconds int64) (float64, error)
 
-	// get function's cold start time
+	// ColdStart provides function's cold start time
 	ColdStart(functionName string, SinceSeconds int64) (float64, error)
 
-	// get function's current CPU and memory usage for each replica
-	TopPods(functionName string) (map[string]int64, map[string]int64, error)
+	// TopPods provides function's current average CPU and memory percentage usage for each replica
+	TopPods(functionName string) (map[string]float64, map[string]float64, error)
 
-	// get nodes current CPU and memory usage
+	// TopNodes provides nodes current CPU and memory percentage usage
 	TopNodes() ([]types.Node, error)
 }
