+++ pkg/types/model.go
@@ -1,15 +1,15 @@
 package types
 
 type Function struct {
-	Name           string           `json:"name"`
-	Replicas       int              `json:"replicas"`
-	InvocationRate float64          `json:"invocation_rate"`
-	ResponseTime   float64          `json:"response_time"`
-	ProcessingTime float64          `json:"processing_time"`
-	Throughput     float64          `json:"throughput"`
-	ColdStart      float64          `json:"cold_start"`
-	Cpu            map[string]int64 `json:"cpu"`
-	Mem            map[string]int64 `json:"mem"`
+	Name           string             `json:"name"`
+	Replicas       int                `json:"replicas"`
+	InvocationRate float64            `json:"invocation_rate"`
+	ResponseTime   float64            `json:"response_time"`
+	ProcessingTime float64            `json:"processing_time"`
+	Throughput     float64            `json:"throughput"`
+	ColdStart      float64            `json:"cold_start"`
+	Cpu            map[string]float64 `json:"cpu"`
+	Mem            map[string]float64 `json:"mem"`
 }
 
 type Node struct {
